data_wrangling/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
Flask-SQLAlchemy==2.3.2